+++ tailwind.config.js
@@ -6,7 +6,6 @@ export default {
     "./src/**/*.{js,ts,jsx,tsx,vue}",
     "./cards/src/**/*.{svelte,js,ts}",
     "./shared/src/**/*.{svelte,js,ts}",
-    "./cards-vue/src/**/*.{js,ts,vue}",
   ],
   theme: {
     extend: {},